Refactor data generation with optional imports and fallback mechanisms
