+++ new_text/data/quality_checked/results_343_quality_checked.json
@@ -1,0 +1,1 @@
+[]
