+++ new_text/data/rewritten/rewritten_qa.json
@@ -1,0 +1,1 @@
+[]
