+++ new_text/data/texts/sample_optics.txt
@@ -1,0 +1,23 @@
+光谱分析是研究物质与电磁辐射相互作用的重要方法：
+
+1. 吸收光谱：
+   - 当光通过物质时，特定波长的光被吸收
+   - 吸收峰的位置和强度反映物质的组成和浓度
+   - Beer-Lambert定律：A = εcl
+
+2. 发射光谱：
+   - 物质受激发后发出特征光谱
+   - 原子发射光谱具有线状特征
+   - 分子发射光谱通常为带状
+
+3. 拉曼光谱：
+   - 基于拉曼散射效应
+   - 提供分子振动和转动信息
+   - 适合研究分子结构和化学键
+
+4. 红外光谱：
+   - 波长范围：2.5-25μm
+   - 反映分子振动模式
+   - 用于化合物鉴定和结构分析
+
+光谱技术在材料科学、化学分析、生物医学等领域有广泛应用。
