+++ new_text/data/texts/sample_semiconductor.txt
@@ -1,0 +1,18 @@
+IGZO（铟镓锌氧化物）是一种新型的氧化物半导体材料，具有以下特点：
+
+1. 高载流子迁移率：IGZO材料的电子迁移率可达10-50 cm²/V·s，远高于非晶硅的迁移率。
+
+2. 良好的均匀性：IGZO薄膜在大面积基板上具有良好的均匀性，适合制造大尺寸显示器。
+
+3. 低温工艺：IGZO可以在相对较低的温度下进行加工，有利于在塑料基板上制造柔性显示器。
+
+4. 透明性：IGZO在可见光范围内具有高透明度，适合制造透明电子器件。
+
+5. 稳定性：IGZO材料具有良好的化学稳定性和电学稳定性。
+
+在TFT-LCD和OLED显示器中，IGZO TFT作为像素开关，具有以下优势：
+- 低功耗：由于高迁移率，可以使用较小的晶体管尺寸
+- 高开关速度：适合高分辨率和高刷新率显示
+- 良好的关态电流特性：有利于提高对比度
+
+IGZO技术的发展推动了高分辨率、低功耗显示技术的进步，特别是在移动设备和大尺寸电视领域有重要应用。
